AI/chat/requirements.txt
fastapi
uvicorn
pymupdf
python-dotenv
sentence-transformers
faiss-cpu
requests
numpy
python-multipart